send metric by metricdata
